+++ packages/design-system/src/components/KBar/KBar.styles.ts
@@ -3,9 +3,6 @@ import { css, darkTheme } from '../../stitches.config'
 /**
  * @styles
  */
-// @note(design-system) uh, this would not happen outside of monorepo
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
 const KBarPositionerStyle = css({
   background: 'rgba(255 255 255 / 0.8)',
   zIndex: '$max',
@@ -13,9 +10,6 @@ const KBarPositionerStyle = css({
     background: 'rgba(0 0 0 / 0.8)',
   },
 })
-// @note(design-system) uh, this would not happen outside of monorepo
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
 const KBarAnimatorStyle = css({
   backgroundColor: '$loContrast',
   boxShadow: '0 0 0 2px $colors$grayA9',
@@ -28,9 +22,6 @@ const KBarAnimatorStyle = css({
     width: '75%',
   },
 })
-// @note(design-system) uh, this would not happen outside of monorepo
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
 const KBarSearchStyle = css({
   backgroundColor: 'transparent',
   border: 'none',
