+++ packages/design-system/src/components/Overlay/Overlay.styles.ts
@@ -1,5 +1,6 @@
 import { css } from '../../stitches.config'
 
+// @todo(design) darkTheme
 const overlayStyles = css({
   backgroundColor: 'rgba(0, 0, 0, .15)',
 })
