+++ packages/design-system/src/components/Sheet/Sheet.tsx
@@ -2,123 +2,21 @@ import * as DialogPrimitive from '@radix-ui/react-dialog'
 import { Cross1Icon } from '@radix-ui/react-icons'
 import * as React from 'react'
 
-import { styled, keyframes, VariantProps, CSS } from '../../stitches.config'
-import { IconButton, overlayStyles } from '../index'
+import { IconButton } from '../IconButton'
 
-const fadeIn = keyframes({
-  from: { opacity: '0' },
-  to: { opacity: '1' },
-})
-
-const fadeOut = keyframes({
-  from: { opacity: '1' },
-  to: { opacity: '0' },
-})
-
-const StyledOverlay = styled(DialogPrimitive.Overlay, overlayStyles, {
-  position: 'fixed',
-  top: 0,
-  right: 0,
-  bottom: 0,
-  left: 0,
-
-  '&[data-state="open"]': {
-    animation: `${fadeIn} 250ms cubic-bezier(0.22, 1, 0.36, 1)`,
-  },
-
-  '&[data-state="closed"]': {
-    animation: `${fadeOut} 250ms cubic-bezier(0.22, 1, 0.36, 1)`,
-  },
-})
-
-type SheetProps = React.ComponentProps<typeof DialogPrimitive.Root>
+import { StyledOverlay, StyledContent, StyledCloseButton } from './Sheet.styles'
+import type { SheetProps, SheetContentProps } from './Sheet.types'
 
 function Sheet({ children, ...props }: SheetProps) {
   return (
     <DialogPrimitive.Root {...props}>
-      <StyledOverlay />
-      {children}
+      <DialogPrimitive.Portal>
+        <StyledOverlay />
+        {children}
+      </DialogPrimitive.Portal>
     </DialogPrimitive.Root>
   )
 }
-
-const slideIn = keyframes({
-  from: { transform: '$$transformValue' },
-  to: { transform: 'translate3d(0,0,0)' },
-})
-
-const slideOut = keyframes({
-  from: { transform: 'translate3d(0,0,0)' },
-  to: { transform: '$$transformValue' },
-})
-
-const StyledContent = styled(DialogPrimitive.Content, {
-  backgroundColor: '$panel',
-  boxShadow: '$colors$shadowLight 0 0 38px -10px, $colors$shadowDark 0 0 35px -15px',
-  position: 'fixed',
-  top: 0,
-  bottom: 0,
-  width: 250,
-
-  // Among other things, prevents text alignment inconsistencies when dialog can't be centered in the viewport evenly.
-  // Affects animated and non-animated dialogs alike.
-  willChange: 'transform',
-
-  // '&:focus': {
-  //   outline: 'none',
-  // },
-
-  '&[data-state="open"]': {
-    animation: `${slideIn} 250ms cubic-bezier(0.22, 1, 0.36, 1)`,
-  },
-
-  '&[data-state="closed"]': {
-    animation: `${slideOut} 250ms cubic-bezier(0.22, 1, 0.36, 1)`,
-  },
-
-  variants: {
-    side: {
-      top: {
-        $$transformValue: 'translate3d(0,-100%,0)',
-        width: '100%',
-        height: 300,
-        bottom: 'auto',
-      },
-      right: {
-        $$transformValue: 'translate3d(100%,0,0)',
-        right: 0,
-      },
-      bottom: {
-        $$transformValue: 'translate3d(0,100%,0)',
-        width: '100%',
-        height: 300,
-        bottom: 0,
-        top: 'auto',
-      },
-      left: {
-        $$transformValue: 'translate3d(-100%,0,0)',
-        left: 0,
-      },
-    },
-  },
-
-  defaultVariants: {
-    side: 'right',
-  },
-})
-
-const StyledCloseButton = styled(DialogPrimitive.Close, {
-  position: 'absolute',
-  top: '$2',
-  right: '$2',
-})
-
-type SheetContentVariants = VariantProps<typeof StyledContent>
-type DialogContentPrimitiveProps = React.ComponentProps<
-  typeof DialogPrimitive.Content
->
-type SheetContentProps = DialogContentPrimitiveProps &
-  SheetContentVariants & { css?: CSS }
 
 const SheetContent = React.forwardRef<
   React.ElementRef<typeof StyledContent>,
@@ -134,16 +32,18 @@ const SheetContent = React.forwardRef<
   </StyledContent>
 ))
 
-const SheetTrigger = DialogPrimitive.Trigger
 const SheetClose = DialogPrimitive.Close
-const SheetTitle = DialogPrimitive.Title
 const SheetDescription = DialogPrimitive.Description
+const SheetPortal = DialogPrimitive.Portal
+const SheetTitle = DialogPrimitive.Title
+const SheetTrigger = DialogPrimitive.Trigger
 
 export {
   Sheet,
-  SheetContent,
-  SheetTrigger,
   SheetClose,
-  SheetTitle,
+  SheetContent,
   SheetDescription,
+  SheetPortal,
+  SheetTitle,
+  SheetTrigger,
 }
