+++ packages/design-system/src/components/Sheet/index.ts
@@ -1,8 +1,11 @@
 export {
   Sheet,
-  SheetContent,
-  SheetTrigger,
   SheetClose,
-  SheetTitle,
+  SheetContent,
   SheetDescription,
+  SheetPortal,
+  SheetTitle,
+  SheetTrigger,
 } from './Sheet'
+
+export { StyledOverlay, StyledContent, StyledCloseButton } from './Sheet.styles'
